+++ autorest/codegen/serializers/__init__.py
@@ -3,10 +3,19 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
+import re
+import logging
 from typing import Dict, List, Optional, Any, Union
 from pathlib import Path
+import black
 from jinja2 import PackageLoader, Environment, FileSystemLoader, StrictUndefined
-from autorest.codegen.models.operation_group import OperationGroup
+
+from autorest.codegen.models.credential_types import (
+    AzureKeyCredentialType,
+    TokenCredentialType,
+)
+from autorest.codegen.models.imports import FileImport, ImportType
+from autorest.codegen.models.operation import OperationBase
 from autorest.codegen.models.request_builder import OverloadedRequestBuilder
 
 from ...jsonrpc import AutorestAPI
@@ -22,6 +31,12 @@ from .operation_groups_serializer import OperationGroupsSerializer
 from .metadata_serializer import MetadataSerializer
 from .request_builders_serializer import RequestBuildersSerializer
 from .patch_serializer import PatchSerializer
+from .sample_serializer import SampleSerializer
+from ..models import BodyParameter
+from .utils import SAMPLE_AAD_ANNOTATION, SAMPLE_KEY_ANNOTATION
+
+_LOGGER = logging.getLogger(__name__)
+_BLACK_MODE = black.Mode()
 
 __all__ = [
     "JinjaSerializer",
@@ -137,7 +152,16 @@ class JinjaSerializer:
                 )
 
         if self.code_model.options["package_mode"]:
-            self._serialize_and_write_package_files(out_path=namespace_path)
+            self._serialize_and_write_package_files(namespace_path)
+
+        if (
+            self.code_model.options["show_operations"]
+            and self.code_model.operation_groups
+            and self.code_model.options["generate_sample"]
+        ):
+            self._serialize_and_write_sample_folder(
+                env=env, namespace_path=namespace_path
+            )
 
     def _serialize_and_write_package_files(self, out_path: Path) -> None:
         def _serialize_and_write_package_files_proc(**kwargs: Any):
@@ -153,7 +177,7 @@ class JinjaSerializer:
                     self._autorestapi.write_file(output_name, render_result)
 
         def _prepare_params() -> Dict[Any, Any]:
-            package_parts = package_name.split("-")[:-1]
+            package_parts = self.code_model.options["package_name"].split("-")[:-1]
             token_cred = isinstance(
                 getattr(self.code_model.credential, "type", None), TokenCredentialType
             )
@@ -180,14 +204,7 @@ class JinjaSerializer:
             params.update(self.code_model.package_dependency)
             return params
 
-        package_name = (
-            self.code_model.options["package_name"]
-            or self.code_model.client.name.lower()
-        )
-        count = package_name.count("-") + 1
-        for _ in range(count):
-            out_path = out_path / Path("..")
-
+        out_path = self._package_root_folder(out_path)
         if self.code_model.options["package_mode"] in ("dataplane", "mgmtplane"):
             env = Environment(
                 loader=PackageLoader("autorest.codegen", "templates"),
@@ -529,3 +546,162 @@ class JinjaSerializer:
         self._autorestapi.write_file(
             namespace_path / Path("_metadata.json"), metadata_serializer.serialize()
         )
+
+    def _package_root_folder(self, namespace_path: Path) -> Path:
+        count = (self.code_model.options.get("package_name") or "").count("-")
+        extra = 2 if self.code_model.options["multiapi"] else 1
+        for _ in range(count + extra):
+            namespace_path = namespace_path / Path("..")
+        return namespace_path
+
+    def _prepare_sample_render_param(self) -> Dict[Any, Any]:
+        # client params
+        credential = ""
+        annotation = ""
+        credential_type = getattr(self.code_model.credential, "type", None)
+        if isinstance(credential_type, TokenCredentialType):
+            credential = "DefaultAzureCredential"
+            annotation = SAMPLE_AAD_ANNOTATION
+        elif isinstance(credential_type, AzureKeyCredentialType):
+            credential = "AzureKeyCredential"
+            annotation = SAMPLE_KEY_ANNOTATION
+
+        addtional_info = (
+            'key=os.getenv("AZURE_KEY")' if credential == "AzureKeyCredential" else ""
+        )
+        special_param = {
+            "credential": f"{credential}({addtional_info})",
+        }
+        params_positional = [
+            p
+            for p in self.code_model.client.parameters.positional
+            if not (p.optional or p.client_default_value)
+        ]
+        client_params = {
+            p.client_name: special_param.get(
+                p.client_name, f'"{p.client_name.upper()}"'
+            )
+            for p in params_positional
+        }
+
+        # imports
+        imports = FileImport()
+        namespace = (self.code_model.options["package_name"] or "").replace(
+            "-", "."
+        ) or self.code_model.namespace
+        imports.add_submodule_import(
+            namespace, self.code_model.client.name, ImportType.THIRDPARTY
+        )
+        if credential:
+            imports.add_submodule_import(
+                "azure.identity", credential, ImportType.THIRDPARTY
+            )
+
+        return {
+            "imports": imports,
+            "client_params": client_params,
+            "annotation": annotation,
+        }
+
+    @staticmethod
+    def _operation_additional(operation: OperationBase[Any]) -> str:
+        lro = ".result()"
+        paging = "\n    response = [item for item in response]"
+        if operation.operation_type == "paging":
+            return "\n    response = [item for item in response]"
+        if operation.operation_type == "lro":
+            return ".result()"
+        if operation.operation_type == "lropaging":
+            return lro + paging
+        return ""
+
+    @staticmethod
+    def _to_lower_camel_case(name: str) -> str:
+        return re.sub(r"_([a-z])", lambda x: x.group(1).upper(), name)
+
+    @staticmethod
+    def _to_snake_case(name: str) -> str:
+        return re.sub(
+            "((?!^)(?<!_)[A-Z][a-z]+|(?<=[a-z0-9])[A-Z])",
+            r"_\1",
+            name.replace("-", "").replace(" ", "_"),
+        ).lower()
+
+    def _serialize_and_write_sample_folder(
+        self, env: Environment, namespace_path: Path
+    ) -> None:
+        out_path = self._package_root_folder(namespace_path) / Path("generated_samples")
+        sample_params = self._prepare_sample_render_param()
+        cls = lambda x: f'"{x}"' if isinstance(x, str) else str(x)
+        failure_info = '"fail to find required param named {%s} in example file {%s}"'
+        # pylint: disable=too-many-nested-blocks
+        for operation_group in self.code_model.operation_groups:
+            if self.code_model.options["multiapi"]:
+                api_version_folder = f"{operation_group.api_versions[0]}/"
+                sample_params["client_params"][
+                    "api_version"
+                ] = f'"{operation_group.api_versions[0]}"'
+            else:
+                api_version_folder = ""
+            for operation in operation_group.operations:
+                samples = operation.yaml_data["samples"]
+                if not samples:
+                    continue
+                params_positional = [
+                    p
+                    for p in operation.parameters.positional
+                    if not p.client_default_value
+                ]
+                operation_result = self._operation_additional(operation)
+                for key, value in samples.items():
+                    # update client parameters if it is defined in example files
+                    for param in sample_params["client_params"].keys():
+                        param_value = value["parameters"].get(
+                            self._to_lower_camel_case(param)
+                        )
+                        if param_value:
+                            sample_params["client_params"][param] = f'"{param_value}"'
+
+                    # prepare method parameters
+                    operation_params = {}
+                    for param in params_positional:
+                        if isinstance(param, BodyParameter):
+                            name = self._to_lower_camel_case(param.client_name)
+                            fake_value = '"can not find valie value"'
+                        else:
+                            name = param.rest_api_name
+                            fake_value = param.client_name.upper()
+
+                        param_value = value["parameters"].get(name)
+                        if param_value or not param.optional:
+                            if not param_value:
+                                # if can't find required param, need to log it
+                                _LOGGER.warning(failure_info, name, key)
+                                param_value = fake_value
+                            operation_params[param.client_name] = cls(param_value)
+
+                    # serialize and output
+                    try:
+                        file_name = self._to_snake_case(key) + ".py"
+                        sample_params["file_name"] = file_name
+                        file_str = SampleSerializer(
+                            code_model=self.code_model,
+                            env=env,
+                            operation_group_name=operation_group.property_name,
+                            operation_name=operation.name,
+                            operation_params=operation_params,
+                            sample_params=sample_params,
+                            operation_result=operation_result,
+                            origin_file=value.get("x-ms-original-file"),
+                        ).serialize()
+                        self._autorestapi.write_file(
+                            out_path / f"{api_version_folder}{file_name}",
+                            black.format_file_contents(
+                                file_str, fast=True, mode=_BLACK_MODE
+                            ),
+                        )
+                    except Exception as e:  # pylint: disable=broad-except
+                        # sample generation shall not block code generation, so just log error
+                        _LOGGER.error(
+                            "error happens when generate sample with {%s}: {%s}", key, e
+                        )